[update] `www.skyperfectv.co.jp` -> `skyperfectv.co.jp`
