+++ sites/skyperfectv.co.jp/README.md
@@ -1,0 +1,28 @@
+# skyperfectv.co.jp
+
+<https://skyperfectv.co.jp>
+
+## Index
+
+- [Index](#index)
+- [Download the guide](#download-the-guide)
+- [Update channel list](#update-channel-list)
+- [Test](#test)
+
+## Download the guide
+
+```sh
+npm run grab -- --site=skyperfectv.co.jp
+```
+
+## Update channel list
+
+```sh
+npm run channels:parse -- --config=./sites/skyperfectv.co.jp/skyperfectv.co.jp.config.js --output=./sites/skyperfectv.co.jp/skyperfectv.co.jp.channels.xml
+```
+
+## Test
+
+```sh
+npm test -- skyperfectv.co.jp
+```
