--- sites/www.skyperfectv.co.jp/README.md
@@ -1,28 +1,0 @@
-# www.skyperfectv.co.jp
-
-<https://www.skyperfectv.co.jp>
-
-## Index
-
-- [Index](#index)
-- [Download the guide](#download-the-guide)
-- [Update channel list](#update-channel-list)
-- [Test](#test)
-
-## Download the guide
-
-```sh
-npm run grab -- --site=www.skyperfectv.co.jp
-```
-
-## Update channel list
-
-```sh
-npm run channels:parse -- --config=./sites/www.skyperfectv.co.jp/www.skyperfectv.co.jp.config.js --output=./sites/www.skyperfectv.co.jp/www.skyperfectv.co.jp.channels.xml
-```
-
-## Test
-
-```sh
-npm test -- www.skyperfectv.co.jp
-```
